+++ ahcer/src/app/services/episode.service.ts
@@ -1,13 +1,13 @@
 import { Injectable } from '@angular/core';
 import firebase from "firebase/compat";
 import OrderByDirection = firebase.firestore.OrderByDirection;
-import {first, from, map, Observable, switchMap} from "rxjs";
-import {Episode} from "../models/episode";
-import {AngularFirestore} from "@angular/fire/compat/firestore";
-import {convertSnaps} from "./data-utils";
-import {UsersService} from "./users.service";
+import { first, from, map, Observable, switchMap } from "rxjs";
+import { Episode } from "../models/episode";
+import { AngularFirestore } from "@angular/fire/compat/firestore";
+import { convertSnaps } from "./data-utils";
+import { UsersService } from "./users.service";
 import Timestamp = firebase.firestore.Timestamp;
-import {Patient} from "../models/patient";
+import { Patient } from "../models/patient";
 
 @Injectable({
   providedIn: 'root'
@@ -15,26 +15,26 @@ import {Patient} from "../models/patient";
 export class EpisodeService {
 
   constructor(private db: AngularFirestore,
-              private user: UsersService) { }
+    private user: UsersService) { }
 
   getLastFiveEpisodesByPatient(patientId: string,
-                       sortOrder: OrderByDirection) : Observable<Episode[]> {
+    sortOrder: OrderByDirection, UserId?: string,): Observable<Episode[]> {
     return this.user.userId$.pipe(
       switchMap(userId =>
-        this.db.collection(`users/${userId}/patients/${patientId}/episodes`,
-        ref => ref.orderBy('startTime', sortOrder).limit(5))
-        .get()),
+        this.db.collection(`users/${UserId ? UserId : userId}/patients/${patientId}/episodes`,
+          ref => ref.orderBy('startTime', sortOrder).limit(5))
+          .get()),
       first(),
       map(snaps => convertSnaps<Episode>(snaps))
     )
   }
 
-  createEpisode(patientId: string, newEpisode: Partial<Episode>): Observable<any> {
+  createEpisode(patientId: string, newEpisode: Partial<Episode>,UserId?:string): Observable<any> {
     let save$: Observable<any>;
 
     save$ = this.user.userId$.pipe(
       switchMap(userId =>
-        from(this.db.collection(`users/${userId}/patients/${patientId}/episodes`).add(newEpisode))
+        from(this.db.collection(`users/${UserId ? UserId : userId}/patients/${patientId}/episodes`).add(newEpisode))
       ),
       first()
     );
@@ -50,35 +50,35 @@ export class EpisodeService {
   }
 
   get20EpisodesByPatient(patientId: string,
-                         sortOrder: OrderByDirection,
-                         lastStartTime?: Timestamp): Observable<Episode[]> {
+    sortOrder: OrderByDirection,
+    lastStartTime?: Timestamp,UserId?:string): Observable<Episode[]> {
     if (lastStartTime) {
       return this.user.userId$.pipe(
-          switchMap(userId =>
-            this.db.collection(`users/${userId}/patients/${patientId}/episodes`,
+        switchMap(userId =>
+          this.db.collection(`users/${ UserId ? UserId : userId}/patients/${patientId}/episodes`,
             ref => ref.orderBy('startTime', sortOrder).startAfter(lastStartTime).limit(21))
             .get()),
-          first(),
-          map(snaps => convertSnaps<Episode>(snaps))
-        )
+        first(),
+        map(snaps => convertSnaps<Episode>(snaps))
+      )
     }
     else {
       return this.user.userId$.pipe(
         switchMap(userId =>
-          this.db.collection(`users/${userId}/patients/${patientId}/episodes`,
-          ref => ref.orderBy('startTime', sortOrder).limit(21))
-          .get()),
-          first(),
-          map(snaps => convertSnaps<Episode>(snaps))
-        )
+          this.db.collection(`users/${UserId ? UserId : userId}/patients/${patientId}/episodes`,
+            ref => ref.orderBy('startTime', sortOrder).limit(21))
+            .get()),
+        first(),
+        map(snaps => convertSnaps<Episode>(snaps))
+      )
     }
   }
 
   getAllEpisodesByPatient(patientId: string,
-                          sortOrder: OrderByDirection) : Observable<Episode[]> {
+    sortOrder: OrderByDirection, UserId?: string): Observable<Episode[]> {
     return this.user.userId$.pipe(
       switchMap(userId =>
-        this.db.collection(`users/${userId}/patients/${patientId}/episodes`,
+        this.db.collection(`users/${UserId ? UserId : userId}/patients/${patientId}/episodes`,
           ref => ref.orderBy('startTime', sortOrder))
           .get()),
       first(),
@@ -86,21 +86,22 @@ export class EpisodeService {
     )
   }
 
-  updateEpisode(patientId: string, episodeId: string, changes: Partial<Patient>): Observable<any> {
+  updateEpisode(patientId: string, episodeId: string, changes: Partial<Patient>, UserId?: string): Observable<any> {
+
     return this.user.userId$.pipe(
       switchMap(userId =>
-        from(this.db.doc(`users/${userId}/patients/${patientId}/episodes/${episodeId}`).update(changes))
+        from(this.db.doc(`users/${UserId ? UserId : userId}/patients/${patientId}/episodes/${episodeId}`).update(changes))
       ),
       first()
     );
   }
 
-  deleteEpisode(patientId: string, episodeId: string): Observable<any> {
+  deleteEpisode(patientId: string, episodeId: string, UserId?: string): Observable<any> {
+    console.log(UserId)
     return this.user.userId$.pipe(
       switchMap(userId => {
-        console.log(userId)
-          return from(this.db.doc(`users/${userId}/patients/${patientId}/episodes/${episodeId}`).delete())
-        }
+        return from(this.db.doc(`users/${UserId ? UserId : userId}/patients/${patientId}/episodes/${episodeId}`).delete())
+      }
       ),
       first()
     );
